v2/pyproject.toml
[project]
name = "morphseg_v2"
version = "0.1.0"
description = "Second-generation pipeline for morphological segmentation with improved prompting and evaluation"
requires-python = ">=3.13"

dependencies = [
    "bitsandbytes>=0.49.2",
    "datasets>=4.5.0",
    "hydra-core>=1.3.2",
    "mlflow>=3.10.0",
    "omegaconf>=2.3.0",
    "peft>=0.18.1",
    "pytorch-lightning>=2.6.1",
    "torch>=2.10.0",
    "transformers>=5.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/morphseg"]

[project.scripts]
do-train = "morphseg.cli:train"
run-inference = "morphseg.cli:inference"

